+++ app/src/main/java/com/drawiin/myheroes/data/network/model/CharacterDtoMapper.kt
@@ -1,4 +1,4 @@
-package com.drawiin.myheroes.network.model
+package com.drawiin.myheroes.data.network.model
 
 import com.drawiin.myheroes.domain.model.character.Character
 import com.drawiin.myheroes.domain.util.DomainMapper
