+++ app/src/main/java/com/drawiin/myheroes/data/network/model/ThumbnailDto.kt
@@ -1,7 +1,6 @@
-package com.drawiin.myheroes.network.model
+package com.drawiin.myheroes.data.network.model
 
 
-import com.drawiin.myheroes.domain.model.character.Thumbnail
 import com.google.gson.annotations.SerializedName
 
 data class ThumbnailDto(
