+++ app/src/main/java/com/drawiin/myheroes/di/RepositoryModule.kt
@@ -1,10 +1,10 @@
 package com.drawiin.myheroes.di
 
 
-import com.drawiin.myheroes.network.model.CharacterDtoMapper
-import com.drawiin.myheroes.network.service.HeroesService
-import com.drawiin.myheroes.repository.DefaultHeroesRepository
-import com.drawiin.myheroes.repository.HeroesRepository
+import com.drawiin.myheroes.data.network.model.CharacterDtoMapper
+import com.drawiin.myheroes.data.network.service.HeroesService
+import com.drawiin.myheroes.data.repository.DefaultHeroesRepository
+import com.drawiin.myheroes.domain.boundarys.HeroesRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
