+++ app/src/main/java/com/drawiin/myheroes/di/UseCaseModule.kt
@@ -1,8 +1,8 @@
 package com.drawiin.myheroes.di
 
 
-import com.drawiin.myheroes.repository.HeroesRepository
-import com.drawiin.myheroes.useCases.GetHeroes
+import com.drawiin.myheroes.domain.boundarys.HeroesRepository
+import com.drawiin.myheroes.domain.interactors.GetHeroes
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
