+++ app/src/main/java/com/drawiin/myheroes/domain/interactors/GetHeroes.kt
@@ -1,7 +1,7 @@
-package com.drawiin.myheroes.useCases
+package com.drawiin.myheroes.domain.interactors
 
 import com.drawiin.myheroes.domain.model.character.Character
-import com.drawiin.myheroes.repository.HeroesRepository
+import com.drawiin.myheroes.domain.boundarys.HeroesRepository
 
 class GetHeroes (
     private val heroesRepository: HeroesRepository
